IPX fix, now Z starts
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@389 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/hardware/IPX.java
@@ -889,7 +889,7 @@ public class IPX extends Module_base {
         }
         for(/*Bitu*/int m=0;m<6;m++) {
             if(localIpxAddr.netnode[m]!=outbuffer[m+0xa])isloopback=false;
-            if(immedAddr[m]!=0xff) islocalbroadcast=false;
+            if(immedAddr[m]!=(byte)0xff) islocalbroadcast=false;
         }
         //LOG_IPX("SEND crc:%2x",packetCRC(&outbuffer[0], packetsize));
         if(!isloopback) {
